+++ app/Controllers/Pelanggan.php
@@ -8,44 +8,59 @@ use App\Models\M_pelanggan;
 class Pelanggan extends BaseController
 {
     protected $mpelanggan;
+    protected $sisi;
 
     public function __construct()
     {
         $this->mpelanggan = new M_pelanggan();
+        $this->sisi = session()->get('level') == 'admin';
     }
+
     public function index()
     {
-        $data = [
-            'active' => 'pelanggan',
-            'open' => 'master',
-            'pelanggan' => $this->mpelanggan->get()
-        ];
-        return view('pages/pelanggan/pelanggan', $data);
+        if ($this->sisi) {
+            $data = [
+                'active' => 'pelanggan',
+                'open' => 'master',
+                'pelanggan' => $this->mpelanggan->get()
+            ];
+            return view('pages/pelanggan/pelanggan', $data);
+        } else {
+            return redirect()->to('auth');
+        }
     }
 
     // TAMBAH
 
     public function tambah()
     {
-        $data = [
-            'active' => 'pelanggan',
-            'open' => 'master',
-            'kode' => $this->mpelanggan->kd_pel(),
-        ];
-        return view('pages/pelanggan/pelanggan_tambah', $data);
+        if ($this->sisi) {
+            $data = [
+                'active' => 'pelanggan',
+                'open' => 'master',
+                'kode' => $this->mpelanggan->kd_pel(),
+            ];
+            return view('pages/pelanggan/pelanggan_tambah', $data);
+        } else {
+            return redirect()->to('auth');
+        }
     }
 
     public function insert()
     {
-        $post = $this->request->getVar();
-        $query = $this->mpelanggan->simpan($post);
+        if ($this->sisi) {
+            $post = $this->request->getVar();
+            $query = $this->mpelanggan->simpan($post);
 
-        if ($query != true) {
-            session()->setFlashdata('success', 'Data Berhasil di tambah');
-            return redirect('pelanggan');
+            if ($query != true) {
+                session()->setFlashdata('success', 'Data Berhasil di tambah');
+                return redirect('pelanggan');
+            } else {
+                session()->setFlashdata('error', 'Gagal tambah data!');
+                return redirect('pelanggan/tambah');
+            }
         } else {
-            session()->setFlashdata('error', 'Gagal tambah data!');
-            return redirect('pelanggan/tambah');
+            return redirect()->to('auth');
         }
     }
 
@@ -54,25 +69,33 @@ class Pelanggan extends BaseController
 
     public function ubah($id)
     {
-        $data = [
-            'active' => 'pelanggan',
-            'open' => 'master',
-            'pelanggan' => $this->mpelanggan->get($id),
-        ];
-        return view('pages/pelanggan/pelanggan_ubah', $data);
+        if ($this->sisi) {
+            $data = [
+                'active' => 'pelanggan',
+                'open' => 'master',
+                'pelanggan' => $this->mpelanggan->get($id),
+            ];
+            return view('pages/pelanggan/pelanggan_ubah', $data);
+        } else {
+            return redirect()->to('/auth');
+        }
     }
 
     public function update($id)
     {
-        $post = $this->request->getVar();
-        $query = $this->mpelanggan->ubah($post);
+        if ($this->sisi) {
+            $post = $this->request->getVar();
+            $query = $this->mpelanggan->ubah($post);
 
-        if ($query == true) {
-            session()->setFlashdata('success', 'Data berhasil diubah');
-            return redirect('pelanggan');
+            if ($query == true) {
+                session()->setFlashdata('success', 'Data berhasil diubah');
+                return redirect('pelanggan');
+            } else {
+                session()->setFlashdata('error', 'Gagal mengubah data');
+                return redirect('pelanggan/ubah/' . $id);
+            }
         } else {
-            session()->setFlashdata('error', 'Gagal mengubah data');
-            return redirect('pelanggan/ubah/' . $id);
+            return redirect()->to('/auth');
         }
     }
 
@@ -80,14 +103,18 @@ class Pelanggan extends BaseController
 
     public function hapus($id)
     {
-        $hapus = $this->mpelanggan->delete($id);
+        if ($this->sisi) {
+            $hapus = $this->mpelanggan->delete($id);
 
-        if ($hapus) {
-            session()->setFlashdata('success', 'Data berhasil hapus');
+            if ($hapus) {
+                session()->setFlashdata('success', 'Data berhasil hapus');
+            } else {
+                session()->setFlashdata('error', 'Gagal menghapus data');
+            }
+
+            return redirect('pelanggan');
         } else {
-            session()->setFlashdata('error', 'Gagal menghapus data');
+            return redirect()->to('/auth');
         }
-
-        return redirect('pelanggan');
     }
 }
